Update of ROS-wip branch to mygym 3.10

+++ myGym/envs/gym_env.py
@@ -182,8 +182,8 @@ class GymEnv(CameraEnv):
             "5-network": {"AGRDW" : AaGaRaDaW},
             "5-network": {"AGFDW" : AaGaFaDaW},
             "5-network": {"AGTDW" : AaGaTaDaW}}
-        
-    
+
+
         scheme = "{}-network".format(str(self.num_networks))
         assert self.unwrapped.reward in reward_classes[scheme].keys(), "Failed to find the right reward class. Check reward_classes in gym_env.py"
         self.task = t.TaskModule(task_type=self.task_type,
@@ -231,7 +231,7 @@ class GymEnv(CameraEnv):
                                transform['position'],  self.p.getQuaternionFromEuler(transform['orientation']),
                                useFixedBase=fixedbase,
                                useMaximalCoordinates=maxcoords)
-    
+
     def _load_static_scene_urdf(self, path, name, fixedbase=True):
         transform = self.workspace_dict[self.workspace]['transform']
         object = env_object.EnvObject(os.path.join(pkg_resources.files("myGym"), os.path.join("envs", path)), transform['position'], self.p.getQuaternionFromEuler(transform['orientation']), pybullet_client=self.p, fixed=fixedbase)
@@ -257,14 +257,14 @@ class GymEnv(CameraEnv):
         if self.obs_space == "dict":
             goaldim = int(self.task.obsdim / 2) if self.task.obsdim % 2 == 0 else int(self.task.obsdim / 3)
             self.observation_space = spaces.Dict(
-                {"observation": spaces.Box(-10.0, high=10., shape=(self.task.obsdim,),dtype = np.float64),
+                {"observation": spaces.Box(-10.0, high=10., shape=(self.task.obsdim,),dtype = np.float32),
                  "achieved_goal": spaces.Box(low=-10., high=10., shape=(goaldim,)),
                  "desired_goal": spaces.Box(low=-10., high=10., shape=(goaldim,))})
         else:
             observationDim = self.task.obsdim
-            observation_high = np.array([100] * observationDim, dtype = np.float64)
+            observation_high = np.array([100] * observationDim)
             self.observation_space = spaces.Box(-observation_high,
-                                                observation_high, dtype = np.float64)
+                                                observation_high)
 
 
     def _set_action_space(self):
@@ -297,7 +297,7 @@ class GymEnv(CameraEnv):
             self.action_low = np.append(self.action_low, np.array(self.robot.gjoints_limits[0]))
             self.action_high = np.append(self.action_high, np.array(self.robot.gjoints_limits[1]))
 
-        self.action_space = spaces.Box(self.action_low, self.action_high, dtype = np.float64)
+        self.action_space = spaces.Box(self.action_low, self.action_high)
 
     def _rescale_action(self, action):
         """
@@ -538,7 +538,7 @@ class GymEnv(CameraEnv):
                 self.human.point_finger_at(position=self.task_objects["goal_state"].get_position())
             self.p.stepSimulation()
         self.episode_steps += 1
-        
+
     def choose_goal_object_by_human_with_keys(self, objects: List[EnvObject]) -> EnvObject:
         self.text_id = self.p.addUserDebugText("Point the human's finger via arrow keys at the goal object and press enter", [1, 0, 0.5], textSize=1)
         move_factor = 10  # times 1 cm
